Create separate worker class for download video

+++ app/src/main/java/com/qs/tv/tvplayer/data/DownloadWorker.kt
@@ -1,47 +1,35 @@
 package com.qs.tv.tvplayer.data
 
 import android.app.DownloadManager
-import android.content.BroadcastReceiver
 import android.content.Context
-import android.content.Intent
 import android.net.Uri
 import androidx.work.Worker
 import androidx.work.WorkerParameters
-import com.qs.tv.tvplayer.room.AppRoomDatabase
-import com.qs.tv.tvplayer.room.entity.PlayerItem
 import com.qs.tv.tvplayer.utils.DownloadData
 import com.qs.tv.tvplayer.utils.FileUtils
+import com.qs.tv.tvplayer.utils.JsonKeys
 import java.io.File
+import java.io.IOException
 
-
-class DownloadDummyImage(private val mContext: Context, workerParams: WorkerParameters) :
-    Worker(mContext, workerParams) {
+class DownloadWorker(private val mContext: Context, workerParams: WorkerParameters) : Worker(mContext, workerParams) {
 
     override fun doWork(): Result {
 
-        //mContext.registerReceiver(onComplete, IntentFilter(DownloadManager.ACTION_DOWNLOAD_COMPLETE))
+        val data = inputData
+        val url: String? = data.getString(JsonKeys.KEY_URL)
+        val path: String? = data.getString(JsonKeys.KEY_PATH)
 
-        for ((index, url) in DummyUrls.images.withIndex()) {
-
-            val name = "Image$index"
-            val path = FileUtils.getImageExternalStorageDirectory(mContext) + name + ".jpg"
+        if (url != null && path != null) {
             val file = File(path)
             if (FileUtils.isExit(path)) {
                 FileUtils.deleteFile(file)
             }
 
-            val item = PlayerItem()
-            item.name = name
-            item.storagePath = path
-            item.url = url
-            item.storageThumbnailPath = path
-            item.thumbnailUrl = ""
-            item.isVideo = false
-            AppRoomDatabase.getInstance(mContext).playerItemDao().insert(item)
-
-            //startDownload(url, file)
-            DownloadData.download(url, file)
-
+            try {
+                DownloadData.download(url, file)
+            } catch (e: IOException) {
+                return Result.failure()
+            }
         }
 
         return Result.success()
@@ -49,7 +37,7 @@ class DownloadDummyImage(private val mContext: Context, workerParams: WorkerPara
 
     private fun startDownload(url: String, file: File) {
         val request: DownloadManager.Request = DownloadManager.Request(Uri.parse(url))
-        request.setDescription("Download image...") // display in android notification
+        request.setDescription("Download video...") // display in android notification
         request.setTitle("tvPlayer") // display in android notification
         //request.setDestinationInExternalFilesDir(mContext, FileUtils.getImagePath(mContext), path)
         //val destinationUri = FileProvider.getUriForFile(mContext, FileUtils.FILE_PROVIDER, File(path))
@@ -62,12 +50,5 @@ class DownloadDummyImage(private val mContext: Context, workerParams: WorkerPara
         * you can also stop the download by using this id
         * */
         val downloadId = manager.enqueue(request)
-    }
-
-    private val onComplete: BroadcastReceiver = object : BroadcastReceiver() {
-
-        override fun onReceive(ctxt: Context?, intent: Intent?) {
-            // your code
-        }
     }
 }
